+++ src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.exceptions.DuplicationException;
 import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.dto.ReservationTimeRequest;
-import roomescape.reservation.dto.ReservationTimeResponse;
+import roomescape.reservation.dto.request.ReservationTimeRequest;
+import roomescape.reservation.dto.response.ReservationTimeResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.theme.domain.Theme;
