+++ src/main/java/roomescape/reservation/service/WaitingService.java
@@ -11,9 +11,9 @@ import roomescape.member.dto.MemberRequest;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Waiting;
-import roomescape.reservation.dto.ReservationOrWaitingResponse;
-import roomescape.reservation.dto.WaitingRequest;
-import roomescape.reservation.dto.WaitingResponse;
+import roomescape.reservation.dto.request.WaitingRequest;
+import roomescape.reservation.dto.response.ReservationOrWaitingResponse;
+import roomescape.reservation.dto.response.WaitingResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.WaitingRepository;
