+++ src/test/java/roomescape/reservation/dto/response/ReservationResponseTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.dto.response;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static roomescape.InitialReservationFixture.RESERVATION_1;
